fix: 회원관리 | JWT token seq int to long 변경 [S07P31D205-46]

See merge request s07-final/S07P31D205!50

+++ Backend/sdutyplus/src/main/java/com/d205/sdutyplus/domain/jwt/support/JwtRequestFilter.java
@@ -44,7 +44,7 @@ public class JwtRequestFilter extends OncePerRequestFilter{
             System.out.println("인증완료");
             String userSeq = JwtUtils.getUserSeq(token);
             //뽑아낸 정보를 인증 정보에 저장
-            Authentication auth = new UsernamePasswordAuthenticationToken(Integer.parseInt(userSeq), null, null);
+            Authentication auth = new UsernamePasswordAuthenticationToken(Long.parseLong(userSeq), null, null);
             SecurityContextHolder.getContext().setAuthentication(auth);
         }
         filterChain.doFilter(request, response);
